+++ packages/mobile_sidebar/lib/common/menu_item.dart
@@ -7,42 +7,37 @@ class MenuItem {
     @required this.title,
     @required this.child,
     @required this.color,
+    this.actions,
+    this.showAppBar = true,
+    this.floatingActionButton,
+    this.floatingActionButtonLocation,
   });
   final String title;
   final String subtitle;
   final IconData icon;
   final Widget child;
   final Color color;
+  final List<Widget> actions;
+  final bool showAppBar;
+  final FloatingActionButton floatingActionButton;
+  final FloatingActionButtonLocation floatingActionButtonLocation;
 
   void push(BuildContext context) => Navigator.push(
         context,
         MaterialPageRoute(
-          builder: (context) => _MobileView(item: this),
-        ),
+            builder: (context) => Scaffold(
+                  appBar: _buildAppBar(context),
+                  body: child,
+                )),
       );
-}
 
-class _MobileView extends StatelessWidget {
-  const _MobileView({
-    Key key,
-    this.appBar = false,
-    @required this.item,
-  }) : super(key: key);
-
-  final MenuItem item;
-  final bool appBar;
-
-  @override
-  Widget build(BuildContext context) {
-    if (appBar) {
-      return Scaffold(
-        appBar: AppBar(
-          title: Text(item.title),
-        ),
-        body: item.child,
-      );
-    }
-
-    return item.child;
+  AppBar _buildAppBar(BuildContext context) {
+    return showAppBar
+        ? AppBar(
+            backgroundColor: Theme.of(context).primaryColor,
+            title: Text(title ?? ''),
+            actions: <Widget>[if (actions != null) ...actions],
+          )
+        : null;
   }
 }