+++ packages/persist_theme/example/lib/main.dart
@@ -1,13 +1,11 @@
 import 'package:flutter/material.dart';
 
 import 'package:persist_theme/persist_theme.dart';
+import 'package:provider/provider.dart';
 import 'package:scoped_model/scoped_model.dart';
 
 import 'dart:io' show Platform;
 import 'package:flutter/foundation.dart';
-
-// The existing imports
-// !! Keep your existing impots here !!
 
 /// main is entry point of Flutter application
 void main() {
@@ -32,34 +30,20 @@ void _setTargetPlatformForDesktop() {
   }
 }
 
-class MyApp extends StatefulWidget {
-  @override
-  _MyAppState createState() => _MyAppState();
-}
+final _model = ThemeModel();
 
-class _MyAppState extends State<MyApp> {
-  final ThemeModel _model = ThemeModel();
-
-  @override
-  void initState() {
-    try {
-      _model.init();
-    } catch (e) {
-      print("Error Loading Theme: $e");
-    }
-    super.initState();
-  }
-
+class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    return ScopedModel<ThemeModel>(
-      model: _model,
-      child: new ScopedModelDescendant<ThemeModel>(
-        builder: (context, child, model) => MaterialApp(
-              theme: model.theme,
-              darkTheme: model.darkTheme,
-              home: HomeScreen(),
-            ),
+    return ListenableProvider<ThemeModel>(
+      builder: (_) => _model..init(),
+      child: Consumer<ThemeModel>(
+        builder: (context, model, child) {
+          return MaterialApp(
+            theme: model.theme,
+            home: HomeScreen(),
+          );
+        },
       ),
     );
   }
@@ -68,7 +52,7 @@ class _MyAppState extends State<MyApp> {
 class HomeScreen extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    final _theme = ScopedModel.of<ThemeModel>(context, rebuildOnChange: true);
+    final _theme = Provider.of<ThemeModel>(context);
     return Scaffold(
       appBar: AppBar(
         title: const Text('Persist Theme'),
@@ -87,19 +71,19 @@ class HomeScreen extends StatelessWidget {
                 Flex(
                   direction: Axis.horizontal,
                   children: <Widget>[
-                    Flexible(child: PrimaryColorPicker(type: PickerType.block)),
-                    Flexible(child: AccentColorPicker(type: PickerType.block)),
+                    Flexible(child: PrimaryColorPicker()),
+                    Flexible(child: AccentColorPicker()),
                   ],
                 ),
-                DarkAccentColorPicker(type: PickerType.block),
+                DarkAccentColorPicker(),
               ]
             : <Widget>[
                 DarkModeSwitch(),
                 TrueBlackSwitch(),
                 CustomThemeSwitch(),
-                PrimaryColorPicker(type: PickerType.block),
-                AccentColorPicker(type: PickerType.block),
-                DarkAccentColorPicker(type: PickerType.block),
+                PrimaryColorPicker(),
+                AccentColorPicker(),
+                DarkAccentColorPicker(),
               ],
       ),
       floatingActionButton: FloatingActionButton(